WIP: implement privileged api
Implement new spire API that allows users to get SVID of other workloads
by passing the selectors of such processes.

The new api is available on the admin socket (the same used for the
debug api). A new "authorized_users_privileged_api" field is added in
the agent configuration to set the IDs of the users that are able to
use this API.

Admin and downstream SVIDs are not sent over this API. (NOTE: this will
be true if that conditional on the code were working, but it's not...)